Move download_article_xml over to the article_processing provider.
